consolidate help/world views
