+++ ishar/apps/core/templates/layout.html
@@ -127,7 +127,7 @@
                         </li>
 {% endif %}
                         <li class="nav-item" title="World">
-                            <a class="bi bi-globe-asia-australia nav-link" href="{% url "world" %}">
+                            <a class="bi bi-globe-asia-australia nav-link" href="{% url "world" %}#world">
                                 World
                             </a>
                         </li>
