+++ ishar/apps/core/urls.py
@@ -1,8 +1,8 @@
 """isharmud.com URL configuration."""
 from django.urls import include, path
 
+from ishar.apps.help.views import WorldView
 from ishar.apps.players.views.who import PlayerWhoView
-from ishar.apps.help.views.world import WorldView
 
 from .views import HistoryView, HomeView, StartView, SupportView
 from .views.auth import IsharLoginView, IsharLogoutView
